Add count to 10
